fix(csv): call `Application.app_dir/2` at runtime
Comparing both calls, the system return the following paths:

* compile time: `/app/_build/prod/lib/growth/priv/indicators`
* runtime: `/app/lib/growth-0.1.0/priv/indicators`

This happens because the project folder structure changes between
compilation/release and the final image. This is due to the fact that
the final `docker` image copies the content from
`/app/_build/${MIX_ENV}/rel/growth` from the `builder` stage into the
`/app` folder in the final image.