fix(csv): call `Application.app_dir/2` at runtime (#13)
Currently, the service is throwing errors because the `ets` tables aren't available.

These errors happen because the `priv` folder changes between the `builder` stage and the final one in the `Dockerfile.

Comparing both calls, the system returns the following paths:

* compile time: `/app/_build/prod/lib/growth/priv/indicators`
* runtime: `/app/lib/growth-0.1.0/priv/indicators`

To fix this issue, the system now gets the reference to the `priv` dir inside the function needing it.