feat: add charts (#21)
* feat: add charts

* chore(calculate): convert `classify` to `add_classification`

Also, reuse `add_percentile`, `add_zscore` and `format_result`
functions.

This changes keep the same pattern previously used and avoid warnings
during compilation.

* chore(calculate): move `percentile` calculation to `Growth.Percentile`

Separate the percentile calculation to it's own module to keep a similar
interface to the `Growth.Zscore` module.

* chore(calculate): move `classify` to `Growth.Classify`

Separate the classification of anthropometric measurements to it's own
module to keep a similar interface to the `Growth.ZScore` module.

* chore(calculate): remove unused match clause

* chore(child): compose methods to calculate age in decimal

* chore(load_reference): fix refactor to fetch reference

* use `get_reference` to fetch reference data
  * adjust gender to `atom` and use it in match clause
* comment unused `ets_table_name` and `interpolate` functions
  * they can be useful for future refactoring

* feat(chart): add module to handle chart data

The module `Growth.LoadReferenceChart` is responsible for loading
reference data for a given measurement type and gender, centered around
a specific age.

It also supports interpolation of data between points to create a
smoother chart.

* chore(chart): fetch data from `Growth.LoadReferenceChart` module

* fix(load): remove commented code

* fix(growth): failing tests

Mainly renaming result keys and some matches.

Also, removed the use of `:meck`.

* fix(loader): ensure proper age unit conversion

* fix(classify): typo in weight classification head

* chore(chart): use the same gender pattern as for data loading

* chore(chart): add telemetry spans for chart data loading

* chore(deps): remove unused dep `:meck`

* chore: document chart loading telemetry events

* feat: resize front end

---------

Co-authored-by: Joao P Dubas <joao.dubas@gmail.com>